Basic 1.20 common port
A rudimentary port of Ecologics to 1.20.

+++ common/src/main/java/samebutdifferent/ecologics/entity/ai/navigation/BetterWallClimberNavigation.java
@@ -50,8 +50,8 @@ public class BetterWallClimberNavigation extends GroundPathNavigation {
                 if (!this.pathToPosition.closerToCenterThan(this.mob.position(),
                         Math.max(this.mob.getBbWidth(), 1.0D))
                         && (!(this.mob.getY() > (double) this.pathToPosition.getY())
-                        || !(new BlockPos(this.pathToPosition.getX(), this.mob.getY(),
-                        this.pathToPosition.getZ())).closerToCenterThan(this.mob.position(),
+                        || !(new BlockPos((int)this.pathToPosition.getX(), (int)this.mob.getY(),
+                        (int)this.pathToPosition.getZ())).closerToCenterThan(this.mob.position(),
                         Math.max(this.mob.getBbWidth(), 1.0D)))) {
                     this.mob.getMoveControl().setWantedPosition(this.pathToPosition.getX(), this.pathToPosition.getY(),
                             this.pathToPosition.getZ(), this.speedModifier);
